[Refactor]: Service에 conversation package 생성
