+++ src/main/java/com/shinhan/knockknock/controller/ConversationController.java
@@ -1,7 +1,7 @@
 package com.shinhan.knockknock.controller;
 
 import com.shinhan.knockknock.domain.dto.conversationroom.ConversationRequest;
-import com.shinhan.knockknock.service.ConversationService;
+import com.shinhan.knockknock.service.conversation.ConversationService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
