+++ src/main/java/com/shinhan/knockknock/controller/ConversationRoomController.java
@@ -4,13 +4,12 @@ import com.shinhan.knockknock.domain.dto.conversationroom.ConversationRoomCreate
 import com.shinhan.knockknock.domain.dto.conversationroom.ConversationRoomResponse;
 import com.shinhan.knockknock.domain.dto.conversationroom.ConversationRoomUpdateRequest;
 import com.shinhan.knockknock.domain.dto.conversationroom.MessageResponse;
-import com.shinhan.knockknock.service.ConversationRoomService;
+import com.shinhan.knockknock.service.conversation.ConversationRoomService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
