+++ src/main/java/com/shinhan/knockknock/domain/dto/conversationroom/ClassificationResponse.java
@@ -1,0 +1,15 @@
+package com.shinhan.knockknock.domain.dto.conversationroom;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ClassificationResponse {
+    String mainTaskNumber;
+    String subTaskNumber;
+}
