+++ src/main/resources/prompts/classification.prompt
@@ -1,7 +1,15 @@
-Your task is to categorize user inputs into one of two categories, each associated with a task number, and return the result in JSON format:
+Your task is to categorize user inputs into one of three main categories (with possible subcategories), each associated with a task number, and return the result in JSON format:
 
-1.Daily Conversation (Task 000): These are casual, non-service related interactions. Examples include greetings, small talk, or general inquiries about the user's day, mood, or feelings.
-2.Welfare Service Guidance (Task 001): These are inquiries related to welfare services, support, or any specific assistance related to the user's well-being. Examples include questions about financial aid, healthcare, or social services.
+Main Categories:
+1.Daily Conversation (Main Task 000): These are casual, non-service related interactions. Examples include greetings, small talk, or general inquiries about the user's day, mood, or feelings.
+2.Welfare Services (Main Task 001): These are inquiries related to welfare services, support, or any specific assistance related to the user's well-being. This category includes:
+ - Service Guidance (Sub Task 001-01): Questions about welfare service availability, registration, and general information.
+3.Financial Services (Main Task 002): These are inquiries related to financial services, such as banking, insurance, or financial advice.
 
-For each user input, determine the appropriate category and return the corresponding task number in the following JSON format:
-`{   "taskNumber": "000"  // or "001" }`
+Response Format:
+For each user input, determine the appropriate main category and subcategory (if applicable) and return the corresponding task numbers in the following JSON format:
+
+{
+  "mainTaskNumber": "000",  // "000", "001", or "002"
+  "subTaskNumber": "001-01"  // Use sub-task numbers if applicable, otherwise null
+}
